# portfolio-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "FiniBoot Portfolio Service API",
        "contact": {}
    },
    "host": "localhost:8003",
    "basePath": "/",
    "paths": {
        "/analytics/coupons": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Считает прогноз выплат по всем облигациям в портфеле",
                "tags": [
                    "analytics"
                ],
                "summary": "Аналитика купонных выплат",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID портфеля",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CouponAnalytics"
                        }
                    }
                }
            }
        },
        "/bond/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portfolio"
                ],
                "summary": "Добавить облигацию в портфель",
                "parameters": [
                    {
                        "description": "Данные для добавления облигации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RequestAddBond"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/bonds": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portfolio"
                ],
                "summary": "Получить состав облигаций в портфеле",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID портфеля",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ResponsePortfolioBond"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portfolio"
                ],
                "summary": "Создать новый портфель",
                "parameters": [
                    {
                        "description": "Название портфеля",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RequestCreatePortfolio"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "portfolio"
                ],
                "summary": "Удалить портфель",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID портфеля",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "Проверяет работоспособность сервиса",
                "tags": [
                    "system"
                ],
                "summary": "Проверка связи",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/portfolios": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portfolio"
                ],
                "summary": "Список всех портфелей пользователя",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ResponsePortfolios"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/transactions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "portfolio"
                ],
                "summary": "История транзакций портфеля",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID портфеля",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PortfolioTransaction"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CouponAnalytics": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "month": {
                    "type": "string"
                }
            }
        },
        "dto.RequestAddBond": {
            "type": "object",
            "properties": {
                "isin": {
                    "type": "string"
                },
                "portfolio_id": {
                    "type": "integer"
                },
                "purchase_date": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "sell_date": {
                    "type": "string"
                }
            }
        },
        "dto.RequestCreatePortfolio": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.ResponsePortfolioBond": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "isin": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nominal": {
                    "type": "number"
                },
                "purchase_date": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "sell_date": {
                    "type": "string"
                }
            }
        },
        "dto.ResponsePortfolios": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.PortfolioTransaction": {
            "type": "object",
            "properties": {
                "bond_isin": {
                    "type": "string"
                },
                "bond_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "portfolio_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "purchase_date": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "sell_date": {
                    "type": "string"
                }
            }
        }
    }
}